payment_service/requirements.txt
alembic==1.2.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2==2.8.3
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.5
Werkzeug==0.16.0
yandex-money-sdk==0.1.3
